ported over hess + musa pages
