+++ src/app/projects/(finance-project-pages)/finance-project-layout.tsx
@@ -32,12 +32,12 @@ const formatMarketCap = (marketCap: number): string => {
 
 const getStockData = (segment: string | null) => {
   switch (segment) {
-    case "hess":
-      return { symbol: "HES", purchasePrice: 108.06 };
-    case "murphy-usa":
-      return { symbol: "MUSA", purchasePrice: 108.06 };
     case "advanced-investments":
       return { symbol: "RBLX", purchasePrice: 108.06 };
+    case "hess":
+      return { symbol: "HES", purchasePrice: 107.63 };
+    case "murphy-usa":
+      return { symbol: "MUSA", purchasePrice: 139.4 };
     default:
       return {
         symbol: "",
@@ -48,7 +48,7 @@ const getStockData = (segment: string | null) => {
 
 const getClassProfileData = (segment: string | null) => {
   switch (segment) {
-    case "hess":
+    case "advanced-investments":
       return {
         heading: "[BUAD 427] Advanced Investments",
         dateString: "Fall 2021",
@@ -61,6 +61,30 @@ const getClassProfileData = (segment: string | null) => {
                     Other financial instruments covered include Eurodollar
                     futures, credit spreads, mortgage loans, mortgage
                     pass-through securities, and interest rate swaps.`,
+      };
+    case "hess":
+      return {
+        heading: "[BUAD 421] Student Managed Investment Fund",
+        dateString: "Spring 2022",
+        description: `This course provides a hands-on experience with portfolio
+                management and security analysis through the management of the
+                Mason School Student Managed Investment Fund (SMIF). Students
+                must select companies from an S&P stock universe, do research on
+                their business model and competitive environment, make forecasts
+                of future financial performance and conduct valuation analyses,
+                write an investment report, and present an oral recommendation
+                to colleagues and faculty for inclusion in a real endowment
+                portfolio of common stocks.`,
+      };
+    case "murphy-usa":
+      return {
+        heading: "[BUAD 329] Corporate Valuation and Credit Analysis",
+        dateString: "Spring 2021",
+        description: `This course focuses on common methodologies for valuing
+                    corporate entities used by professionals working in
+                    investments, private equity, venture capital and investment
+                    banking. It aims to familiarize students with various data
+                    sources and software used in the financial industry.`,
       };
     default:
       return {
@@ -136,13 +160,13 @@ const FinanceLayout = ({ children }: FinanceLayoutProps) => {
   }, [isLoading, errors.length, quoteData, companyData, symbol]);
 
   return (
-    <div className="md:max-w-xl lg:max-w-3xl xl:max-w-5xl 2xl:max-w-6xl flex flex-col items-center space-y-10 lg:space-y-14 mx-10 md:mx-auto">
+    <main className="md:max-w-xl lg:max-w-3xl xl:max-w-5xl 2xl:max-w-6xl flex flex-col items-center space-y-10 lg:space-y-14 mx-10 md:mx-auto">
       <ClassProfile heading={heading} dateString={dateString}>
         {description}
       </ClassProfile>
       <StockCard data={stockCardData} errors={errors} loading={isLoading} purchasePrice={purchasePrice} />
       {children}
-    </div>
+    </main>
   );
 };
