+++ src/app/projects/(finance-project-pages)/hess/page.tsx
@@ -1,0 +1,11 @@
+import { Metadata } from "next";
+
+import HessPage from "./hess-page";
+
+export const metadata: Metadata = {
+  title: "HES Pitch",
+};
+
+export default async function Page() {
+  return <HessPage />;
+}
