+++ src/app/projects/(finance-project-pages)/murphy-usa/page.tsx
@@ -1,0 +1,11 @@
+import { Metadata } from "next";
+
+import MurphyUSAPage from "./murphy-usa-page";
+
+export const metadata: Metadata = {
+  title: "MUSA Credit Analysis",
+};
+
+export default async function Page() {
+  return <MurphyUSAPage />;
+}
